refactor🧹 : myPosition과 mapPosition 코드 분리
