+++ pages/index.tsx
@@ -1,24 +1,23 @@
 import type { NextPage } from 'next'
 import Head from 'next/head'
 import { useEffect, useState } from 'react'
-import { useRecoilState, useSetRecoilState } from 'recoil';
+import { useRecoilState, useRecoilValue } from 'recoil';
 import SignalList from '../Components/SignalList';
-import myPositionState from '../recoil/myPosition/atom';
-import { aroundSignalsAtom, mapAroundSignalsAtom } from '../recoil/aroundSignals/atom';
 import filterSignals from '../utils/filterSignals';
-import mapPositionAtom from '../recoil/mapPosition/atom';
 import getDistance from '../utils/getDistance';
-import mapMovingAtom from '../recoil/mapMoving/atom';
 import getSignalData from '../utils/getSignalData';
 import placeSignal from '../utils/placeSignal';
 import removeSignals from '../utils/removeSignal';
+import myPositionState from '../recoil/myPosition/atom';
+import mapPositionAtom from '../recoil/mapPosition/atom';
+import { mapAroundSignalsAtom } from '../recoil/aroundSignals/atom';
+import mapMovingAtom from '../recoil/mapMoving/atom';
 
 const Home: NextPage = () => {
-  const [myPosition, setMyPosition] = useRecoilState(myPositionState);
+  const myPosition = useRecoilValue(myPositionState);
   const [mapPosition, setMapPosition] = useRecoilState(mapPositionAtom);
   const [isMapMoving, setIsMapMoving] = useRecoilState(mapMovingAtom);
   const [mapAroundSignals, setMapAroundSignals] = useRecoilState(mapAroundSignalsAtom);
-  const setAroundSignals = useSetRecoilState(aroundSignalsAtom);
   const [mapLoaded, setMapLoaded] = useState<boolean>(false);
   const [map, setMap] = useState();
 
@@ -58,18 +57,6 @@ const Home: NextPage = () => {
   }, [isMapMoving, mapPosition]);
 
   useEffect(() => {
-    (async() => {
-      const response = await getSignalData(myPosition);
-
-      if (!response.length) return;
-  
-      const filteredSignals = filterSignals(response);
-  
-      setAroundSignals(filteredSignals as any);
-    })();
-  }, [myPosition]);
-
-  useEffect(() => {
     if (!mapLoaded) return;
 
     new window.kakao.maps.load(() => {
@@ -82,10 +69,10 @@ const Home: NextPage = () => {
 
       map.panTo(new window.kakao.maps.LatLng(myPosition.lat, myPosition.lng));
 
-      const gpsContent = `<div id="me"></div>`;
+      const myIcon = `<div id="me"></div>`;
       const currentOverlay = new window.kakao.maps.CustomOverlay({
           position: new window.kakao.maps.LatLng(myPosition.lat, myPosition.lng),
-          content: gpsContent,
+          content: myIcon,
           map: map
       });
 
@@ -105,25 +92,6 @@ const Home: NextPage = () => {
   }, [mapLoaded, myPosition.lat, myPosition.lng]);
 
   useEffect(() => {
-    navigator.geolocation.getCurrentPosition((pos) => {
-      const cord = pos.coords;
-      // const newPosition = {
-      //   lat: cord.latitude,
-      //   lng: cord.longitude,
-      // };
-      const newPosition = {
-        lat: 37.57814842135318,
-        lng: 126.88837721721241,
-      };
-      //dmc position
-
-      setMyPosition(newPosition);
-    }, (err) => {
-      console.log(err);
-    });
-  }, []);
-
-  useEffect(() => {
     const script = document.createElement("script");
 
     script.src = `//dapi.kakao.com/v2/maps/sdk.js?appkey=${process.env.NEXT_PUBLIC_KAKAOMAP_API_KEY}&autoload=false`;
@@ -136,7 +104,8 @@ const Home: NextPage = () => {
     <div>
       <Head>
         <title>Walk</title>
-        <meta name="description" content="Generated by create next app" />
+        <meta name="description" content="Walk Efficiently using Traffic light information" />
+        <meta name="keyword" content="Walk, Traffic Light" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
